use Tabs for generator-modification-form
