+++ src/components/dialogs/network-modifications/generator/modification/generatorTabs.utils.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+export enum GeneratorDialogTab {
+    CONNECTIVITY_TAB = 0,
+    SETPOINTS_AND_LIMITS_TAB = 1,
+    SPECIFIC_TAB = 2,
+    ADDITIONAL_INFORMATION_TAB = 3,
+}
